+++ .storybook/webpack.config.js
@@ -18,14 +18,6 @@ module.exports = {
       },
     ],
   },
-  entry: {
-    manager: [require.resolve('../../manager')],
-  },
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: 'static/[name].bundle.js',
-    publicPath: '/',
-  },
   plugins: [
     new OccurenceOrderPlugin(),
     new webpack.HotModuleReplacementPlugin(),