+++ src/lib/components/consignments/__stories__/index.ts
@@ -1,4 +1,5 @@
 import "./consignments.story"
-
-import "./consignments-todo.story"
 import "./typography.story"
+
+import "./consignments-search.story"
+import "./consignments-todo.story"